+++ HISTORY.rst
@@ -3,6 +3,11 @@
 History
 =======
 
+1.2.0 (2019-10-11)
+---------------------
+* Add the capability to run in quiet mode when running locally on a workstation
+  so the job stdout and stderr are written to log files only.
+
 1.1.0 (2019-06-07)
 ---------------------
 * HPC array job command lines are quoted and executed in a subshell by default with better support for complex command lines.